+++ src/main/java/com/sparta/server/threeserving/order_management/dto/response/DashboardTrendResponse.java
@@ -3,7 +3,6 @@ package com.sparta.server.threeserving.order_management.dto.response;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
@@ -25,8 +24,8 @@ public class DashboardTrendResponse {
         private Long rejectedCount;
         private Long canceledCount;
         private Long totalSalesAmount;
-        private BigDecimal avgCookTime;
-        private BigDecimal acceptRate;
+        private Double avgCookTime;
+        private Double acceptRate;
 
     }
 }
